Rework the cleanup process to allow for a more gentle approach and avoid

See merge request exascale/external/jupiter/preps!6

+++ libexec/preps/os/cleanup_residual_user_processes.sh
@@ -12,17 +12,23 @@
 preps::os::cleanup_residual_user_processes() {
   main::log_event -level "TRACE" -message "Entering Module: [${FUNCNAME[0]}]"
   if (( SLURM_JOB_UID > 1000 )); then
-    # 1st Pass
-    if killall --signal SIGKILL --user "${SLURM_JOB_USER}"; then
-      main::log_event -level "INFO" -message "Killed Residual User Processes for User: [${SLURM_JOB_USER}]"
+    # 1st gentle pass
+    if killall --signal SIGTERM --user "${SLURM_JOB_USER}"; then
+      main::log_event -level "INFO" -message "Killed Residual User Processes for User: [${SLURM_JOB_USER}] via SIGTERM"
     fi
-    # 2nd Pass
-    local pid
-    while read -r -u 9 pid; do
-      if kill --signal SIGKILL "${pid}"; then
-        main::log_event -level "INFO" -message "Killed Residual User Process: [${pid}]"
+    # 2nd not so gentle pass
+    if pgrep -u "${SLURM_JOB_USER}" > /dev/null; then
+      # Some processes did not finish yet, give them the chance for 2 seconds
+      sleep 2
+      # Can't wait anymore, kill them
+      if killall --signal SIGKILL --user "${SLURM_JOB_USER}"; then
+        main::log_event -level "INFO" -message "Lingering processes after gentle termination: Killed Residual User Processes for User: [${SLURM_JOB_USER}] via SIGKILL"
       fi
-    done 9< <(pgrep -u "${SLURM_JOB_USER}" 2>/dev/null | xargs)
+    fi
+    # If not all processes were successfully killed, report it
+    if pgrep -u "${SLURM_JOB_USER}" > /dev/null; then
+      main::log_event -level "INFO" -message "SIGTERM and SIGKILL did not kill all processes of user ${SLURM_JOB_USER} in time. Hoping health checks get that."
+    fi
   fi
   main::log_event -level "TRACE" -message "Exiting Module: [${FUNCNAME[0]}]"
   return 0
